Avoid deprecation error in the use of imageio.imread
As of imageIO 2.16.0 (Feb22) there are now a v2 and v3 namespaces in
addition to the top-level namespace. In order to avoid a deprecation
warning, the call to imageio.imread has been replaced by imageio.v2.imread.

The version of imageio in requirements.txt has also been changed to the
most recent version of the module (2.22.2).